Fixed a 32/64 bit compile error with memset.
git-svn-id: http://svn.bramp.net:8080/svn/PhD/Work/userspace/threadnetperf@2723 e02701c3-087c-644a-a2a3-551e05f81084

+++ client_thread.c
@@ -214,8 +214,7 @@ void* client_thread(void *data) {
 		goto cleanup;
 	}
 
-	memset( buffer, (int)BUFFER_FILL, settings.message_size ); // TODO fix 32bit linux compile problem
-
+	memset( buffer, BUFFER_FILL, settings.message_size );
 #ifdef USE_EPOLL
 	readFD_epoll = epoll_create(clients);
 	if(readFD_epoll == -1) {
